requirements.txt
fpdf==1.7.2
Pillow==9.4.0
Pillow==10.0.0
streamlit==1.19.0
